front: move login screen to its own route (#1152)
