+++ public/mobile-app/src/lib/france-connect.ts
@@ -25,7 +25,7 @@ export const franceConnectLogout = async (
   redirect_url: string | null = null
 ) => {
   const redirect_uri = encodeURIComponent(
-    redirect_url || `${PUBLIC_APP_URL}/?is_logged_out`
+    redirect_url || `${PUBLIC_APP_URL}/?is_logged_out#/login`
   );
   let post_logout_redirect_uri = redirect_uri;
   if (PUBLIC_FC_PROXY_BASE_URL) {
