+++ public/mobile-app/src/routes/login/page.svelte.test.ts
@@ -1,0 +1,182 @@
+import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
+import '@testing-library/jest-dom/vitest';
+import { render, screen, waitFor } from '@testing-library/svelte';
+import * as navigationMethods from '$app/navigation';
+import { toastStore } from '$lib/state/toast.svelte';
+import { userStore } from '$lib/state/User.svelte';
+import { mockUserInfo } from '$tests/utils';
+import Page from './+page.svelte';
+
+describe('/+page.svelte', () => {
+  let originalWindow: typeof globalThis.window;
+
+  beforeEach(() => {
+    originalWindow = globalThis.window;
+    userStore.connected = null;
+  });
+
+  afterEach(() => {
+    globalThis.window = originalWindow;
+    vi.resetAllMocks();
+  });
+
+  test('should navigate to homepage when user has already logged in', async () => {
+    await userStore.login(mockUserInfo);
+
+    const spy = vi
+      .spyOn(navigationMethods, 'goto')
+      .mockImplementation(() => Promise.resolve());
+
+    render(Page);
+    expect(spy).toHaveBeenCalledWith('/');
+  });
+
+  test('should render FranceConnect button', async () => {
+    // Given
+    vi.spyOn(globalThis, 'fetch').mockResolvedValue(
+      new Response(JSON.stringify(mockUserInfo), { status: 200 })
+    );
+
+    // When
+    render(Page);
+
+    // Then
+    await waitFor(() => {
+      const franceConnectButton = screen.getByRole('button', {
+        name: 'S’identifier avec FranceConnect',
+      });
+      expect(franceConnectButton).toHaveTextContent('S’identifier avec FranceConnect');
+    });
+  });
+
+  test('should display network-error page on FranceConnect login button when back is down', async () => {
+    // Given
+    vi.spyOn(globalThis, 'fetch').mockRejectedValue(new Error());
+    const spy = vi
+      .spyOn(navigationMethods, 'goto')
+      .mockImplementation(() => Promise.resolve());
+
+    render(Page);
+    await waitFor(() => {
+      const franceConnectLoginButton = screen.getByRole('button', {
+        name: 'S’identifier avec FranceConnect',
+      });
+
+      // When
+      franceConnectLoginButton.click();
+
+      // Then
+      expect(spy).toHaveBeenCalledWith('/#/network-error');
+    });
+  });
+
+  test('should call authorize endpoint when click on FranceConnect login button', async () => {
+    // Given
+    vi.spyOn(globalThis, 'fetch').mockResolvedValue(new Response('', { status: 200 }));
+    vi.stubGlobal('location', { href: 'fake-link' });
+
+    render(Page);
+    await waitFor(() => {
+      const franceConnectLoginButton = screen.getByRole('button', {
+        name: 'S’identifier avec FranceConnect',
+      });
+
+      // When
+      franceConnectLoginButton.click();
+
+      // Then
+      expect(globalThis.window.location.href).toContain('login-france-connect');
+    });
+  });
+
+  test('should display an error message if login failed', async () => {
+    // Given
+    const { page } = await import('$app/state');
+    const mockSearchParams = new URLSearchParams(
+      'error=some error message&error_description=some error description'
+    );
+    vi.spyOn(page.url, 'searchParams', 'get').mockReturnValue(mockSearchParams);
+
+    render(Page);
+
+    // Then
+    const errorMessage = await screen.findByText('some error message');
+    expect(errorMessage).toBeInTheDocument();
+    const errorDescription = await screen.findByText('some error description');
+    expect(errorDescription).toBeInTheDocument();
+  });
+
+  test('should not display any error message if the user aborted the connection', async () => {
+    // Given
+    const { page } = await import('$app/state');
+    const mockSearchParams = new URLSearchParams(
+      'error=access_denied&error_description=User auth aborted'
+    );
+    vi.spyOn(page.url, 'searchParams', 'get').mockReturnValue(mockSearchParams);
+
+    render(Page);
+
+    // Then
+    const errorMessage = await screen.queryByText('access_denied');
+    expect(errorMessage).toBeNull();
+    const errorDescription = await screen.queryByText('User auth aborted');
+    expect(errorDescription).toBeNull();
+  });
+
+  test('should logout the app if an error is about FranceConnect', async () => {
+    // Given
+    const { page } = await import('$app/state');
+    const mockSearchParams = new URLSearchParams(
+      'error=some error message&error_type=FranceConnect'
+    );
+    vi.spyOn(page.url, 'searchParams', 'get').mockReturnValue(mockSearchParams);
+
+    render(Page);
+
+    // Then
+    const errorMessage = await screen.findByText('some error message');
+    expect(errorMessage).toBeInTheDocument();
+    expect(window.localStorage.getItem('access_token')).toEqual(null);
+  });
+
+  test('should add toast when user is logged out', async () => {
+    // Given
+    const { page } = await import('$app/state');
+    const mockSearchParams = new URLSearchParams('is_logged_out');
+    vi.spyOn(page.url, 'searchParams', 'get').mockReturnValue(mockSearchParams);
+
+    const spy = vi.spyOn(toastStore, 'addToast');
+
+    // When
+    render(Page);
+
+    // Then
+    await waitFor(async () => {
+      expect(spy).toHaveBeenCalledWith(
+        'Vous avez bien été déconnecté(e)',
+        'success',
+        3000,
+        false
+      );
+    });
+  });
+
+  test('should display connection help links', async () => {
+    HTMLDialogElement.prototype.showModal = vi.fn();
+    HTMLDialogElement.prototype.close = vi.fn();
+    HTMLDialogElement.prototype.show = vi.fn();
+
+    // When
+    render(Page);
+
+    // Then
+    const connectionHelpButton = screen.getByTestId('connection-help-button');
+    expect(screen.queryByTestId('connection-help-link-url')).not.toBeInTheDocument();
+    await waitFor(() => {
+      connectionHelpButton.click();
+      // now the popup is open
+      expect(screen.queryByTestId('connection-help-link-url')).toBeInTheDocument();
+      expect(screen.queryByTestId('connection-help-link-email')).toBeInTheDocument();
+    });
+  });
+});
